// src/backend/interfaces.tsx
export interface FormattedWorklog {
  'Avatar':string;
  'Author ID': string;
  'Author Name': string;
  'User Link': string;
  'Department':string;
  'Department Ledger Code': string;
  'Log ID': number;
  'Account ID': string;
  'Account Name': string;
  'Issue Name': string;
  'Issue ID': string;
  'Issue Link': string;
  'Logged Date': string;
  'Logged Time': string;
  'Created At': string;
  'Updated At': string;
  'Time Spent': number;
  'Billable': number;
}

export interface Worklogs {
  Worklog: Worklog;
}

export interface RootTimeLogs {
  user: {
    User: User;
  };
  worklogs: Worklogs[];
}

export interface AvatarUrls {
  '48x48': string;
  '24x24': string;
  '16x16': string;
  '32x32': string;
}

export interface User {
  self: string;
  accountId: string;
  accountType: string;
  emailAddress?: string;
  avatarUrls: AvatarUrls;
  displayName: string;
  active: boolean;
  locale: string;
}
// ...
export interface Worklog {
  self: string;
  tempoWorklogId: number;
  issue: {
    self: string;
    id: number;
    issueName: string;
  };
  accountId: string;
  accountName: string;
  timeSpentSeconds: number;
  billableSeconds: number;
  issueStartDate: string;
  loggedTime: string;
  loggedDate: string;
  createdAt: string;
  updatedAt: string;
  author: {
    avatar: string;
    self: string;
    userID: string;
    displayName: string;
  };
  department: string;
  departmentLedgerCode: string;
}
export interface IssueBase {
  self: string;
  id: string;
}

export interface JiraResponseBase {
  self: string;
  metadata: Metadata;
}

export interface WorklogDictionary {
  userId: string;
  displayName: string;
  worklogs: FilteredWorklog[];
}
export interface Author {
  self: string;
  accountId: string;
}
// ...
export interface Issue extends IssueBase {
  expand: string;
  key: string;
  fields: Fields;
}

export interface FormattedIssue {
  issueName: string;
  id: number;
  self: string;
}

export interface Fields {
  parent?: Parent;
  customfield_13580: unknown; // Replace 'any' with more specific types as needed
  customfield_13460: unknown;
  // Include other fields as required
  customfield_13418: CustomField[];
  customfield_13317: CustomField[];
  customfield_13316: CustomField[];
  // Add other custom fields as needed
  issuetype: IssueType;
  project: Project;
  assignee: Assignee;
  reporter: Reporter;
  worklog: Worklog;
  status: Status;
  // Add other properties as required
}